[Configuration] Add more fine-grained permissions to configurations (#438)
* [Configuration] Add more fine-gained permissions to configurations - resolves #357

* Apply php-cs-fixer changes

* [Configuration] Add more fine-gained permissions to configurations - fix when no permissions exists resolves #357

* [Configuration] Add more fine-gained permissions to configurations - disable config button instead of hide #357

* [Configuration] Add more fine-gained permissions to configurations - permissions save data refactoring #357

* [Configuration] Add more fine-grained permissions to configurations - related to https://github.com/pimcore/data-hub/pull/438

* Apply suggestions from code review


* Update src/Configuration/Permission.php


* Apply suggestions from code review


* Update src/Migrations/PimcoreX/Version20211108160248.php


* [Configuration] Add more fine-grained permissions to configurations

* Apply php-cs-fixer changes

* [Configuration] Add more fine-grained permissions to configurations

* [Configuration] Add more fine-grained permissions to configurations

* Apply php-cs-fixer changes

* Fine grained permissions refactoring (#453)

* refactored permissions

* Apply php-cs-fixer changes

* use user/role name instead of id for checking permissions

* Apply php-cs-fixer changes

* [Configuration] Add more fine-grained permissions to configurations

* [Configuration] Add more fine-grained permissions to configurations

* Apply php-cs-fixer changes

* [Configuration] Add more fine-grained permissions to configurations

* added category for permissions

* Apply php-cs-fixer changes

* added missing translation

+++ src/Configuration.php
@@ -17,6 +17,8 @@ namespace Pimcore\Bundle\DataHubBundle;
 
 use Pimcore\Bundle\DataHubBundle\Event\ConfigurationEvents;
 use Pimcore\Model\AbstractModel;
+use Pimcore\Model\User;
+use Pimcore\Tool\Admin;
 use Symfony\Component\EventDispatcher\GenericEvent;
 
 /**
@@ -250,6 +252,10 @@ class Configuration extends AbstractModel
      */
     public function save(): void
     {
+        if (!$this->isAllowed('update')) {
+            throw new \Exception('Permissions missing to save the configuration');
+        }
+
         $event = new GenericEvent($this);
         $event->setArgument('configuration', $this);
         \Pimcore::getEventDispatcher()->dispatch($event, ConfigurationEvents::CONFIGURATION_PRE_SAVE);
@@ -261,6 +267,11 @@ class Configuration extends AbstractModel
 
         if (!isset($this->configuration['workspaces'])) {
             $this->configuration['workspaces'] = [];
+        }
+
+        if (!isset($this->configuration['permissions'])) {
+            $this->configuration['permissions']['user'] = [];
+            $this->configuration['permissions']['role'] = [];
         }
 
         if (isset($this->configuration['general']['writeable'])) {
@@ -300,6 +311,10 @@ class Configuration extends AbstractModel
      */
     public function delete(): void
     {
+        if (!$this->isAllowed('delete')) {
+            throw new \Exception('Permissions missing to delete the configuration');
+        }
+
         $this->getDao()->delete();
 
         $event = new GenericEvent($this);
@@ -419,11 +434,78 @@ class Configuration extends AbstractModel
         return $this->configuration['security'] ?? [];
     }
 
+    /**
+     * @return mixed
+     */
+    public function getPermissionsConfig()
+    {
+        return $this->configuration['permissions'] ?? [];
+    }
+
     public function __clone()
     {
         if ($this->dao) {
             $this->dao = clone $this->dao;
             $this->dao->setModel($this);
         }
+    }
+
+    /**
+     * @internal
+     *
+     * @param string $type
+     * @param ?User $user
+     *
+     * @return bool
+     */
+    public function isAllowed(string $type, ?User $user = null)
+    {
+        if (null === $user) {
+            $user = Admin::getCurrentUser();
+        }
+
+        if (!$user) {
+            if (php_sapi_name() === 'cli') {
+                return true;
+            }
+
+            return false;
+        }
+
+        //everything is allowed for admin
+        if ($user->isAdmin() || $user->isAllowed('plugin_datahub_admin')) {
+            return true;
+        }
+
+        /**
+         * If there are no specific permissions set for this configuration then check the global adapter permission.
+         * Otherwise, check the configured permissions only and ignore the global adapter permission.
+         */
+        $configKey = 'plugin_datahub_adapter_' . $this->getType();
+
+        $permissionConfig = $this->getPermissionsConfig();
+        $permissionSets = [];
+        foreach ($permissionConfig['user'] ?? [] as $userConfig) {
+            $permissionSets[$userConfig['name']] = $userConfig;
+        }
+        foreach ($permissionConfig['role'] ?? [] as $roleConfig) {
+            $permissionSets[$roleConfig['name']] = $roleConfig;
+        }
+
+        if (empty($permissionSets)) {
+            return $user->isAllowed($configKey);
+        } else {
+            if (isset($permissionSets[$user->getName()])) {
+                return $permissionSets[$user->getName()][$type] ?? false;
+            }
+            foreach ($user->getRoles() as $roleId) {
+                $role = User\Role::getById($roleId);
+                if (isset($permissionSets[$role->getName()][$type]) && $permissionSets[$role->getName()][$type] === true) {
+                    return $permissionSets[$role->getName()][$type];
+                }
+            }
+        }
+
+        return false;
     }
 }
